# docs/swagger.yml
swagger: '2.0'
info:
  version: 1.0.0
  title: XDCMaster APIs
  description: Happy to code XDCMaster APIs
  license:
    name: Github
    url: 'https://github.com/XinFin/XDCmaster'
...
schemes:
  - http
  - https
consumes:
  - application/json
produces:
  - application/json
paths:
...
  '/api/voters/{voter}/candidates':
    parameters:
      - name: voter
        in: path
        type: string
        description: voter's address
        required: true
      - name: limit
        in: query
        description: Number of record in a query
        type: number
        required: false
      - name: page
        in: query
        description: Page number
        type: number
        required: false
    get:
      tags:
        - Voters
      summary: Get voted candidates
      responses:
        '200':
          description: OK
          schema:
            $ref: '#/definitions/voterCandidates'
        '406':
          description: Not Acceptable
  '/api/voters/{voter}/rewards':
    parameters:
      - name: voter
        in: path
        type: string
        description: candidate's address
        required: true
      - name: limit
        in: query
        type: number
        description: number of records
        required: false
      - name: page
        in: query
        type: number
        description: page number
        required: false
    get:
      tags:
        - Voters
      summary: Get reward of voter
      responses:
        '200':
          description: OK
          schema:
            $ref: '#/definitions/voterRewards'
        '406':
          description: Not Acceptable
...
definitions:
...
  voterCandidates:
    properties:
      items:
        type: array
        description: Candidate's data
        items:
          type: object
          description: candidate's information
          example:
            candidate: '0xfc5571921c6d3672e13b58ea23dea534f2b35fa0'
            capacity: '10000000000000000000'
            capacityNumber: 10
            candidateName: Earth
      total:
        type: number
        description: Total number of rewards
        example: 100
  voterRewards:
    properties:
      items:
        type: array
        description: Voter's reward array
        items:
          type: object
          description: voter's reward
          example:
            _id: 5c19847f0e77307940be922a
            epoch: 3193
            startBlock: 2872801
            endBlock: 2873700
            address: '0x11621900588eca4410c00097a9f59237f34064cd'
            validator: '0x11621900588eca4410c00097a9f59237f34064cd'
            reason: MasterNode
            lockBalance: '1000'
            reward: '25.97042513863216266174'
            rewardTime: '2018-12-18T23:36:18.000Z'
            signNumber: 843
            __v: 0
            createdAt: '2018-12-18T23:36:31.580Z'
            updatedAt: '2018-12-18T23:36:31.580Z'
            candidateName: '0x11621900588eca4410c00097a9f59237f34064cd'
      total:
        type: number
        description: Number of candidate
        example: 100
...